requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
psycopg2-binary